CardioScanX/requirements.txt
streamlit
torch
numpy
pandas
plotly
matplotlib
scipy
wfdb
reportlab
Pillow
scikit-learn
